Support for image files and view its contents
Image files can be viewed as an image now. Before it could only be
viewed as binary code in the Grace editor. In order to achieve this, a
new section with id="image-view" has been added to the HTML which
contains the image display. A new function called isImage is in files.js
to identify which files are images based on their name.