Drag and drop files and directories
Allow files and directories to be dragged around between
directories. Before the positions of files and directories
were fixed once they were uploaded or created under directoris.
Now the positions of files and directories can be changed as
they can be dragged around between directories. Once the
position of a directory has been changed, the files under that
directories will also be changed.