Tidy names and generate Minigrace version number
The title of the editor page no longer includes the experimental suffix.
The full Minigrace script no longer has 'min' in its extension, as the
file is not typically minified. The version number of the compiler is
now included on the editor page, generated from a field in the compiler.
Styling for the objectdraw height-calculator is also included.