Add npm install and eslint configuration files
The editor should now be easier to setup with npm. The configuration
file specifies browserify as a dependency, and uses it to compile the
scripts when `npm install` is invoked.

An eslint configuration file has been added, and all of the scripts have
been brought in line with the linting rules. The rule no-alert should be
considered temporary, as the alert, confirm, and prompt calls should be
removed in the future.