Allows sidebar to be viewed when zooming in on editor
Modifies CSS (and sidebar.js functions controlling sidebar hiding) to make the sidebar stay when the window is re-sized or the page is zoomed in on. This allows the editor to be zoomed up to 250% in Chrome while still maintaining a semi-reasonable layout.  The user can still hide the sidebar if they wish.