Adds Directory Renaming
This commit adds directory renaming by simulating the addition of a
directory with the new name, dragging-and-dropping contents to the
new directory, and then deleting the old directory (now empty). This will
be revised in the future to be more efficient. Also, adds a button to the
context menu UI for this function.